s/appspot/cloud function/
Before this change, our REST calls to yr.no were made through a Java
proxy running serverless on appspot.

With this change in place, that Java code has now been replaced with a
Google Cloud Function.

Since we have so few users, we get a lot of cold starts, and my hope is
that the cold start performance of a Python Cloud Function is better
than a Java app server.

So hopefully this change will improve worst-case forecast download
performance.