Stop hanging at "Downloading weather..."
By clearing the download progress message after the download is
complete.

The issue used to be that we only downloaded weather data when the
progress message was undefined, to make sure we didn't do multiple
downloads at once.

However, since we never cleared the progress message when not
downloading, we would never download again after the first time.